requirements.txt
pytorch
torchvision
pytorch-ignite
Pillow